LayoutTests/fast/shadow-dom/ShadowRoot-mode.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests support for ShadowRoot.mode attribute.");

var element1 = document.createElement('div');
var shadowRoot1 = element1.attachShadow({ mode: "open" });
shouldBeEqualToString("shadowRoot1.mode", "open");

var element2 = document.createElement('div');
var shadowRoot2 = element2.attachShadow({ mode: "closed" });
shouldBeEqualToString("shadowRoot2.mode", "closed");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
