LayoutTests/fast/shadow-dom/ShadowRoot-mode-expected.txt

Tests support for ShadowRoot.mode attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot1.mode is "open"
PASS shadowRoot2.mode is "closed"
PASS successfullyParsed is true

TEST COMPLETE
